Delete old indexing code
