+++ sql/analyzer/analyzer_test.go
@@ -32,7 +32,7 @@ func TestAddRule(t *testing.T) {
 
 	defRulesCount := countRules(NewDefault(nil).Batches)
 
-	a := NewBuilder(nil).AddPostAnalyzeRule(-1, generateIndexScans).Build()
+	a := NewBuilder(nil).AddPostAnalyzeRule(-1, pushFilters).Build()
 
 	require.Equal(countRules(a.Batches), defRulesCount+1)
 }
@@ -42,7 +42,7 @@ func TestAddPreValidationRule(t *testing.T) {
 
 	defRulesCount := countRules(NewDefault(nil).Batches)
 
-	a := NewBuilder(nil).AddPreValidationRule(-1, generateIndexScans).Build()
+	a := NewBuilder(nil).AddPreValidationRule(-1, pushFilters).Build()
 
 	require.Equal(countRules(a.Batches), defRulesCount+1)
 }
@@ -52,7 +52,7 @@ func TestAddPostValidationRule(t *testing.T) {
 
 	defRulesCount := countRules(NewDefault(nil).Batches)
 
-	a := NewBuilder(nil).AddPostValidationRule(-1, generateIndexScans).Build()
+	a := NewBuilder(nil).AddPostValidationRule(-1, pushFilters).Build()
 
 	require.Equal(countRules(a.Batches), defRulesCount+1)
 }
